Don't check the workflows count
